refactor(sub_agent): extract chaining functions to SubAgent.Chaining
Move then!/3, then/3, and validate_chain_keys!/2 to a dedicated module.
This reduces SubAgent from 1281 to 1164 lines (~117 lines) while keeping
the public API unchanged via defdelegate.
